timer

+++ AS2023Env/BackgroundService.cs
@@ -8,8 +8,9 @@ public class BackgroundService : IHostedService
     private readonly IStorage<Employee> _employeeStorage;
     private readonly ILogger<BackgroundService> _logger;
 
-    private Timer _timer = null;
-    
+    private Timer _fireEmployeeTimer = null;
+    private Timer _resetStatusesTimer = null;
+
     public BackgroundService(IStorage<StaffUnit> staffUnitStorage, IStorage<Employee> employeeStorage, ILogger<BackgroundService> logger)
     {
         _staffUnitStorage = staffUnitStorage;
@@ -19,9 +20,28 @@ public class BackgroundService : IHostedService
 
     public Task StartAsync(CancellationToken cancellationToken)
     {
-        _timer = new Timer(FireEmployee, null, Constants.FireEmployeeDelay, Constants.FireEmployeeDelay);
+        _fireEmployeeTimer = new Timer(FireEmployee, null, Constants.FireEmployeeDelay, Constants.FireEmployeeDelay);
+        _resetStatusesTimer = new Timer(ResetStatuses, null, TimeSpan.Zero, TimeSpan.FromSeconds(1));
         _logger.LogInformation("Фоновый сервис стартовал");
         return Task.CompletedTask;
+    }
+
+    private async void ResetStatuses(object state)
+    {
+        List<StaffUnit> pendingStaffUnits = await _staffUnitStorage
+            .GetList(s => s.Status == StaffUnitStatus.Pending && s.CloseTime != null && s.CloseTime <= DateTime.Now);
+
+        foreach (StaffUnit staffUnit in pendingStaffUnits)
+        {
+            staffUnit.SetClosed();
+            _logger.LogInformation(
+                "Для штатной единицы id = {Id} установлен статус {Status}",
+                staffUnit.Id,
+                staffUnit.Status.ToString()
+            );
+
+            await _staffUnitStorage.Update(staffUnit);
+        }
     }
 
     private async void FireEmployee(object _)
@@ -43,10 +63,13 @@ public class BackgroundService : IHostedService
 
         Employee pickToFire = employees[index];
         StaffUnit staffUnit = (await _staffUnitStorage.GetList(u => u.EmployeeId == pickToFire.Id)).FirstOrDefault();
+        if (staffUnit?.Status == StaffUnitStatus.Pending)
+        {
+            return;
+        }
         if (staffUnit != null)
         {
-            staffUnit.EmployeeId = null;
-            staffUnit.Status = StaffUnitStatus.Opened;
+            staffUnit.SetEmployee(null);
             await _staffUnitStorage.Update(staffUnit);
         }
         await _employeeStorage.Delete(pickToFire.Id);
@@ -56,7 +79,8 @@ public class BackgroundService : IHostedService
 
     public Task StopAsync(CancellationToken cancellationToken)
     {
-        _timer.Change(Timeout.Infinite, 0);
+        _fireEmployeeTimer.Change(Timeout.Infinite, 0);
+        _resetStatusesTimer.Change(Timeout.Infinite, 0);
         _logger.LogInformation("Фоновый сервис останавливается");
         return Task.CompletedTask;
     }
